Update to scrape.py
Works with new SCPD page, asks for the URL instead of the course name, works with new two-factor authentication, doesn't store password in plain-text, does ten at once, and doesn't bother with encoding (just downloads the webm version, it's a nice small version of the video. Download VLC if you can't view them).

Liable to become outdated quickly, as everything is hardcoded in.